+++ internal/provider/provider_sdk.go
@@ -40,6 +40,7 @@ func Provider(v string) *schema.Provider {
 			"scalr_agent_pool":               dataSourceScalrAgentPool(),
 			"scalr_current_account":          dataSourceScalrCurrentAccount(),
 			"scalr_current_run":              dataSourceScalrCurrentRun(),
+			"scalr_current_environment":      dataSourceScalrCurrentEnvironment(),
 			"scalr_iam_team":                 dataSourceScalrIamTeam(),
 			"scalr_iam_user":                 dataSourceScalrIamUser(),
 			"scalr_module_version":           dataSourceModuleVersion(),