+++ CHANGELOG.md
@@ -7,6 +7,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+### Added
+
+- **New resource:** `scalr_drift_detection` ([#473](https://github.com/Scalr/terraform-provider-scalr/pull/473))
+
+
+## [3.10.0] - 2025-10-23
+
+### Changed
+
+- `scalr_provider_configuration`: Service account email is no longer required when using OIDC credentials in the Google provider configuration. ([#469](https://github.com/Scalr/terraform-provider-scalr/pull/469))
+- `scalr_access_policy`: changed type of `role_ids` attribute from TypeList to TypeSet. ([#466](https://github.com/Scalr/terraform-provider-scalr/pull/466))
+- `scalr_iam_team`: changed type of `users` attribute from TypeList to TypeSet. ([#466](https://github.com/Scalr/terraform-provider-scalr/pull/466))
+- `scalr_policy_group`: changed type of `environments` attribute from TypeList to TypeSet. ([#466](https://github.com/Scalr/terraform-provider-scalr/pull/466))
+- `scalr_webhook`: changed type of `events` attribute from TypeList to TypeSet. ([#466](https://github.com/Scalr/terraform-provider-scalr/pull/466))
+- Added sorting for computed attributes with List of String type in resources and data sources. ([#466](https://github.com/Scalr/terraform-provider-scalr/pull/466))
+
+### Required
+
+- scalr-server >= v2510.17.0
+
 ## [3.9.0] - 2025-09-19
 
 ### Changed
@@ -1115,7 +1136,8 @@ Requires Scalr 8.0.1-beta.20200625 at least
 
 - Initial release.
 
-[Unreleased]: https://github.com/Scalr/terraform-provider-scalr/compare/v3.9.0...HEAD
+[Unreleased]: https://github.com/Scalr/terraform-provider-scalr/compare/v3.10.0...HEAD
+[3.10.0]: https://github.com/Scalr/terraform-provider-scalr/releases/tag/v3.10.0
 [3.9.0]: https://github.com/Scalr/terraform-provider-scalr/releases/tag/v3.9.0
 [3.8.0]: https://github.com/Scalr/terraform-provider-scalr/releases/tag/v3.8.0
 [3.7.1]: https://github.com/Scalr/terraform-provider-scalr/releases/tag/v3.7.1